requirements.txt
dj_database_url==1.3.0
Django==5.1.5
django-allauth==65.11.1
django-jsoneditor==0.2.4
djangorestframework==3.16.0
joblib==1.5.0
numpy==2.2.4
opencv-contrib-python==4.12.0.88
opencv-python==4.11.0.86
pandas==2.3.3
Pillow==12.1.0
python-dotenv==1.2.1
reportlab==4.4.7
spacy==3.8.7
ultralytics==8.3.223
